+++ static/src/javascripts/bootstraps/enhanced/common.js
@@ -60,6 +60,7 @@ import {
     consentManagementPlatformUi,
     addPrivacySettingsLink,
 } from 'common/modules/ui/cmp-ui';
+import { signInGate } from 'common/modules/identity/sign-in-gate';
 
 const initialiseTopNavItems = (): void => {
     const header: ?HTMLElement = document.getElementById('header');
@@ -313,6 +314,7 @@ const initialiseBanner = (): void => {
         smartAppBanner,
         adFreeBanner,
         emailSignInBanner,
+        signInGate,
     ];
 
     initBannerPicker(bannerList);
